web-platform/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
sqlmodel>=0.0.16
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.1
python-multipart>=0.0.9
aiofiles>=23.2.1
Pillow>=10.0.0
numpy>=1.24.0
python-pptx>=0.6.21
passlib>=1.7.4
pyjwt[crypto]>=2.8.0
python-jose[cryptography]>=3.3.0
redis>=5.0.1
httpx>=0.26.0
python-dotenv>=1.0.0
